Do not check whether the PDB exists
just `createOrPatch` it.

While being here:
* Deleted the AppWorkload controller unit test and moved the integration
  test next to the controller
* Refactored the `PDBUpdater` unit test into an integration test

Issue: https://github.com/cloudfoundry/korifi/issues/1721
