+++ statefulset-runner/controllers/appworkload_to_stset_test.go
@@ -10,6 +10,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/client-go/kubernetes/scheme"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 )
@@ -22,7 +23,6 @@ var _ = Describe("AppWorkload to StatefulSet Converter", func() {
 	)
 
 	BeforeEach(func() {
-		Expect(korifiv1alpha1.AddToScheme(scheme.Scheme)).To(Succeed())
 		appWorkload = createAppWorkload("some-namespace", "guid_1234")
 		converter = controllers.NewAppWorkloadToStatefulsetConverter(zap.New(zap.WriteTo(GinkgoWriter), zap.UseDevMode(true)), scheme.Scheme)
 	})
@@ -258,3 +258,68 @@ var _ = Describe("AppWorkload to StatefulSet Converter", func() {
 		})
 	})
 })
+
+func expectedValFrom(fieldPath string) *corev1.EnvVarSource {
+	return &corev1.EnvVarSource{
+		FieldRef: &corev1.ObjectFieldSelector{
+			APIVersion: "",
+			FieldPath:  fieldPath,
+		},
+	}
+}
+
+func createAppWorkload(namespace, name string) *korifiv1alpha1.AppWorkload {
+	return &korifiv1alpha1.AppWorkload{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
+			Namespace: namespace,
+		},
+		Spec: korifiv1alpha1.AppWorkloadSpec{
+			AppGUID:          "premium_app_guid_1234",
+			GUID:             "guid_1234",
+			Version:          "version_1234",
+			Image:            "gcr.io/foo/bar",
+			ImagePullSecrets: []corev1.LocalObjectReference{{Name: "some-secret-name"}},
+			Command: []string{
+				"/bin/sh",
+				"-c",
+				"while true; do echo hello; sleep 10;done",
+			},
+			ProcessType: "worker",
+			Env:         []corev1.EnvVar{},
+			StartupProbe: &corev1.Probe{
+				ProbeHandler: corev1.ProbeHandler{
+					HTTPGet: &corev1.HTTPGetAction{
+						Path: "/healthz",
+						Port: intstr.IntOrString{Type: intstr.Int, IntVal: int32(8080)},
+					},
+				},
+				FailureThreshold: 30,
+				PeriodSeconds:    2,
+			},
+			LivenessProbe: &corev1.Probe{
+				ProbeHandler: corev1.ProbeHandler{
+					HTTPGet: &corev1.HTTPGetAction{
+						Path: "/healthz",
+						Port: intstr.IntOrString{Type: intstr.Int, IntVal: int32(8080)},
+					},
+				},
+				PeriodSeconds:    30,
+				FailureThreshold: 1,
+			},
+			Ports:      []int32{8888, 9999},
+			Instances:  1,
+			RunnerName: "statefulset-runner",
+			Resources: corev1.ResourceRequirements{
+				Limits: corev1.ResourceList{
+					corev1.ResourceEphemeralStorage: resource.MustParse("2048Mi"),
+					corev1.ResourceMemory:           resource.MustParse("1024Mi"),
+				},
+				Requests: corev1.ResourceList{
+					corev1.ResourceCPU:    resource.MustParse("5m"),
+					corev1.ResourceMemory: resource.MustParse("1024Mi"),
+				},
+			},
+		},
+	}
+}
