+++ statefulset-runner/controllers/suite_test.go
@@ -1,88 +1,135 @@
+/*
+Copyright 2022.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package controllers_test
 
 import (
+	"context"
+	"path/filepath"
 	"testing"
-
-	"k8s.io/apimachinery/pkg/util/intstr"
+	"time"
 
 	korifiv1alpha1 "code.cloudfoundry.org/korifi/controllers/api/v1alpha1"
-	"code.cloudfoundry.org/korifi/statefulset-runner/fake"
+	. "code.cloudfoundry.org/korifi/statefulset-runner/controllers"
 
+	"github.com/google/uuid"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes/scheme"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/envtest"
+	logf "sigs.k8s.io/controller-runtime/pkg/log"
+	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+	//+kubebuilder:scaffold:imports
+)
+
+// These tests use Ginkgo (BDD-style Go testing framework). Refer to
+// http://onsi.github.io/ginkgo/ to learn more about Ginkgo.
+
+var (
+	cancel    context.CancelFunc
+	k8sClient client.Client
+	testEnv   *envtest.Environment
 )
 
 func TestAppWorkloadsController(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "Controller Suite")
+
+	SetDefaultEventuallyTimeout(10 * time.Second)
+	SetDefaultEventuallyPollingInterval(200 * time.Millisecond)
+
+	RunSpecs(t, "Controller Integration Suite")
 }
 
-var (
-	fakeClient       *fake.Client
-	fakeStatusWriter *fake.StatusWriter
-)
+var _ = BeforeSuite(func() {
+	logf.SetLogger(zap.New(zap.WriteTo(GinkgoWriter), zap.UseDevMode(true)))
 
-var _ = BeforeEach(func() {
-	fakeClient = new(fake.Client)
-	fakeStatusWriter = &fake.StatusWriter{}
-	fakeClient.StatusReturns(fakeStatusWriter)
+	ctx, cancelFunc := context.WithCancel(context.TODO())
+	cancel = cancelFunc
+
+	By("bootstrapping test environment")
+	testEnv = &envtest.Environment{
+		CRDDirectoryPaths: []string{
+			filepath.Join("..", "..", "helm", "korifi", "controllers", "crds"),
+		},
+		ErrorIfCRDPathMissing: true,
+	}
+
+	cfg, err := testEnv.Start()
+	Expect(err).NotTo(HaveOccurred())
+	Expect(cfg).NotTo(BeNil())
+
+	err = korifiv1alpha1.AddToScheme(scheme.Scheme)
+	Expect(err).NotTo(HaveOccurred())
+
+	//+kubebuilder:scaffold:scheme
+
+	webhookInstallOptions := &testEnv.WebhookInstallOptions
+	k8sManager, err := ctrl.NewManager(cfg, ctrl.Options{
+		Scheme:             scheme.Scheme,
+		Host:               webhookInstallOptions.LocalServingHost,
+		Port:               webhookInstallOptions.LocalServingPort,
+		CertDir:            webhookInstallOptions.LocalServingCertDir,
+		LeaderElection:     false,
+		MetricsBindAddress: "0",
+	})
+	Expect(err).NotTo(HaveOccurred())
+
+	logger := ctrl.Log.WithName("statefulset-runner").WithName("AppWorkload")
+	appWorkloadReconciler := NewAppWorkloadReconciler(
+		k8sManager.GetClient(),
+		k8sManager.GetScheme(),
+		NewAppWorkloadToStatefulsetConverter(logger, k8sManager.GetScheme()),
+		NewPDBUpdater(k8sManager.GetClient()),
+		logger,
+	)
+	err = (appWorkloadReconciler).SetupWithManager(k8sManager)
+	Expect(err).NotTo(HaveOccurred())
+
+	k8sClient, err = client.New(cfg, client.Options{Scheme: scheme.Scheme})
+	Expect(err).NotTo(HaveOccurred())
+	Expect(k8sClient).NotTo(BeNil())
+
+	go func() {
+		defer GinkgoRecover()
+		err = k8sManager.Start(ctx)
+		Expect(err).NotTo(HaveOccurred())
+	}()
 })
 
-func createAppWorkload(namespace, name string) *korifiv1alpha1.AppWorkload {
-	return &korifiv1alpha1.AppWorkload{
+var _ = AfterSuite(func() {
+	cancel()
+	By("tearing down the test environment")
+	Expect(testEnv.Stop()).To(Succeed())
+})
+
+func prefixedGUID(prefix string) string {
+	return prefix + "-" + uuid.NewString()[:8]
+}
+
+func createNamespace(ctx context.Context, k8sClient client.Client, name string) *corev1.Namespace {
+	ns := &corev1.Namespace{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: korifiv1alpha1.AppWorkloadSpec{
-			AppGUID:          "premium_app_guid_1234",
-			GUID:             "guid_1234",
-			Version:          "version_1234",
-			Image:            "gcr.io/foo/bar",
-			ImagePullSecrets: []corev1.LocalObjectReference{{Name: "some-secret-name"}},
-			Command: []string{
-				"/bin/sh",
-				"-c",
-				"while true; do echo hello; sleep 10;done",
-			},
-			ProcessType: "worker",
-			Env:         []corev1.EnvVar{},
-			StartupProbe: &corev1.Probe{
-				ProbeHandler: corev1.ProbeHandler{
-					HTTPGet: &corev1.HTTPGetAction{
-						Path: "/healthz",
-						Port: intstr.IntOrString{Type: intstr.Int, IntVal: int32(8080)},
-					},
-				},
-				FailureThreshold: 30,
-				PeriodSeconds:    2,
-			},
-			LivenessProbe: &corev1.Probe{
-				ProbeHandler: corev1.ProbeHandler{
-					HTTPGet: &corev1.HTTPGetAction{
-						Path: "/healthz",
-						Port: intstr.IntOrString{Type: intstr.Int, IntVal: int32(8080)},
-					},
-				},
-				PeriodSeconds:    30,
-				FailureThreshold: 1,
-			},
-			Ports:      []int32{8888, 9999},
-			Instances:  1,
-			RunnerName: "statefulset-runner",
-			Resources: corev1.ResourceRequirements{
-				Limits: corev1.ResourceList{
-					corev1.ResourceEphemeralStorage: resource.MustParse("2048Mi"),
-					corev1.ResourceMemory:           resource.MustParse("1024Mi"),
-				},
-				Requests: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("5m"),
-					corev1.ResourceMemory: resource.MustParse("1024Mi"),
-				},
-			},
+			Name: name,
 		},
 	}
+	Expect(
+		k8sClient.Create(ctx, ns)).To(Succeed())
+	return ns
 }
